test(js): hoist shared fakeSpan/makeTracer into tests/helpers/fakeTracer.ts
