+++ packages/js/tests/helpers/index.ts
@@ -1,1 +1,2 @@
 export * from './FakeApi';
+export * from './fakeTracer';
